Bug fixed
Previously failed to realize that the key for the DataSource was the key in the table - 1 (because the DataSource starts from 0 but I populate the database from 1).

It now works to do a load from the key of the first visible item for a number of pages just as big to cover all the items displayed.